Add Yank plugin for windows.
Since a win32yank shows high cpu usage when yank and paste lot of
info and sometimes didn't yank at all. this plugin is an alternative
to yank and paste by using it as a provider.
Works well and do not break the system by boost CPU usage to 100%

Also change some snippets and identing.